requirements.txt
# ── Core agent ──────────────────────────────────
moshi                        # PersonaPlex / full-duplex voice model
websockets>=12.0             # WebSocket client for PersonaPlex server
sounddevice>=0.4.6           # Microphone input / speaker output

# ── AI / LLM ────────────────────────────────────
anthropic>=0.25.0            # Claude API — booking extraction + RAG

# ── RAG ─────────────────────────────────────────
chromadb>=0.5.0              # Vector store (persisted, local)
sentence-transformers>=2.7.0 # Local embedding model (all-MiniLM-L6-v2)

# ── Document loaders ────────────────────────────
pypdf>=4.0.0                 # Read .pdf files
python-docx>=1.1.0           # Read .docx files

# ── Data / utils ────────────────────────────────
numpy>=1.26.0                # Audio array processing
